requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
Faker==37.3.0
fonttools==4.58.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.2.6
openai==1.84.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pydantic==2.11.5
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
